[process] Emit error from exit when code != 0
Some process error handler was not called when process exited on error.
(`BaseLanguageServerContribution.onDidFailSpawnProcess`).

This is because the process did start, but once it began parsing its
command line arguments, it decided to exit with a return code != 0.

At this point, it is hard to distinguish between a spawn error (which
kind of happens really early) and the program just exiting at some later
point in time.

This commit adds a new method to override for
`LanguageServerContributions` in order to be notified when the process
does exit after successfully starting.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>